+++ lib/dnd/index.ts
@@ -1,6 +1,5 @@
-export { useDragSource } from './useDragSource';
+export { useDragSource, DragPreview } from './useDragSource';
 export { useDropTarget } from './useDropTarget';
-export { DragPreview } from './DragPreview';
 export { useDndStore } from './store';
 export * from './types';
 export * from './utils';
